(Issue #5) Cleaned up FileLogger test significantly.
This revealed that level error checking in the log function can be simplified,
which aligns level error management with that done in StandardLogger and the
_log_debug function in create_log_functions.

Since StandardLogger has a default logger level, we should do the same with the
same default value for the FileLogger.

All related tox tasks passing.  For all tests that confirm correct error
detection, I printed out the actual error messages to confirm that the error
messages are logged correctly and that the tests are catching the intended
errors.