(Issue #10) Change log_debug level argument.
Based on the experience of writing the first developer example, passing the
actual debug level values to log_debug() is better than using shifted levels
that are zero-based.  I found it complicated the documentation in the sphinx
docs and in the example.  Also, the test is now arguably simpler.

Tests passing locally.