(Issue #15) Revert previous break & break tests.
The previous break did cause both the sdist and wheel install subjobs to fail.
The former reported that the file could not be found as expected.  The latter
repored that

ERROR: Invalid requirement: 'poptus-*-py3-none-any.wl': Expected end or
semicolon (after name and no valid version specifier)

Since this is tied to the name, I do believe that the correct failure was
detected.  I suspect that in this case, an error check failed since the wheel
has an invalid file extension.  In other words, this failed before the code
could find out that the file doesn't exist.  Some local testing does indeed
suggest that the wheel code does some upfront checks.  For example with the
correct extension I changed "poptus" to "pptus" and it complains that the file's
name is inconsistent with the package name as specified in the wheel internally.